🔨 Refactor: Model 수정및 폴더 변경
Closes #25
