+++ Bbik/Resource/krdata.json
@@ -4,7 +4,7 @@
         "menus" : [
             { "image": "https://drive.google.com/uc?export=view&id=1vIp5I0Ui5-WDge9FKa0hWM62GrQxi77B", "name": "오색 무지개 바람떡", "price": 4700,"stock": 30, "sales": 25 },
             { "image": "https://drive.google.com/uc?export=view&id=1THiMUJj6k7wSGPFkilX6NPNxPWD42sbk", "name": "불고기 도시락", "price": 5200, "stock": 30, "sales": 18 },
-            { "image": " https://drive.google.com/uc?export=view&id=1dT7JxcOBtW-lgAmFs2xXGmyx-X12KrOs", "name": "제육 도시락", "price": 5800, "stock": 30, "sales": 30 },
+            { "image": "https://drive.google.com/uc?export=view&id=1dT7JxcOBtW-lgAmFs2xXGmyx-X12KrOs", "name": "제육 도시락", "price": 5800, "stock": 30, "sales": 30 },
             { "image": "https://drive.google.com/uc?export=view&id=1KDuJpI9pXgHEfGSDkKxJI-rP4QddCYtO", "name": "치킨마요", "price": 4600, "stock": 30, "sales": 7 },
             { "image": "https://drive.google.com/uc?export=view&id=1xaI8Y__uwefybUHlpRWAbkZdWqEyJfJl", "name": "참치마요", "price": 4500, "stock": 30, "sales": 22 },
             { "image": "https://drive.google.com/uc?export=view&id=17qc-4kv0vv-eC0pql9LcBS_AjNI-ng8i", "name": "돈까스 도시락", "price": 4900, "stock": 30, "sales": 13 },
